Fix Merge. Rename and move template. Remove comments
