+++ aidants_connect_web/templates/aidants_connect_web/statistiques.html
@@ -23,7 +23,7 @@
                 {{ stat_item.title }}{% if stat_item.subtitle %}*{% endif %}
               </h3>
               <h3>
-                <b>{{ stat_item.value }}</b>
+                <strong>{{ stat_item.value }}</strong>
               </h3>
             </div>
             <p class="tile-subtitle">
@@ -47,7 +47,7 @@
                 {{ stat_item.title }}
               </h3>
               <h3>
-                <b>{{ stat_item.value }}</b>
+                <strong>{{ stat_item.value }}</strong>
               </h3>
             </div>
           </div>
