+++ aidants_connect_web/templates/layouts/mandat/mandat_template_20200123.html
@@ -12,13 +12,14 @@
     {% endif %}
   </div>
   
-  <h1>Mandat type d’un aidant réalisant des démarches administratives en ligne au bénéfice d’un usager via le service
+  <h1>Mandat pour réaliser des démarches administratives en ligne au bénéfice d’un usager via le service
     « Aidants Connect »</h1>
 
-  <p>Je soussigné(e), <strong>{{ usager }}</strong> autorise <strong>{{ aidant }}</strong>, exerçant la profession de
-    <strong>{{profession}}</strong> au sein de <strong>{{ organisation }}</strong> à
+  <p>
+    Je soussigné(e), <strong>{{ usager }}</strong> autorise <strong>{{ organisation }}</strong>, dont seuls les agents habilités peuvent faire à la place de, à
     réaliser en mon nom conformément aux dispositions des articles 1984 et suivants du Code civil, les démarches
-    administratives en ligne suivantes:</p>
+    administratives en ligne suivantes, étant précisé qu’il s’agisse de démarches accessibles via FranceConnect ou non :
+  </p>
 
   <ul>
     {% for demarche in demarches %}
@@ -33,9 +34,8 @@
   <p>
     À cette fin :
     <ul>
-      <li><strong>{{ usager }}</strong> autorise <strong>{{ aidant }}</strong> à utiliser ses données à caractère personnel.
-      </li>
-      <li><strong>{{ aidant }}</strong> a rappelé à <strong>{{ usager }}</strong> l’objet de l’intervention, la raison pour
+      <li><strong>{{ usager }}</strong> autorise <strong>{{ organisation }}</strong> à utiliser ses données à caractère personnel.</li>
+      <li><strong>{L'aidant habilité de {{ organisation }}</strong> a rappelé à <strong>{{ usager }}</strong> l’objet de l’intervention, la raison pour
         laquelle ses informations sont collectées et leur utilité ; l’existence de droits
         sur ses données (accès, rectification, suppression, etc.) et la possibilité pour le mandant de retirer à tout
         moment son consentement.</li>
@@ -45,37 +45,34 @@
   <p>
     Le mandataire s’engage à :
     <ol>
-      <li>ne collecter et enregistrer que les seules informations strictement nécessaires au regard des démarches
-        susvisées ;</li>
-      <li>n’utiliser les informations concernant le mandant que pour les seules démarches susvisées. S’il a besoin de
-        les utiliser pour d’autres démarches, il doit au préalable en informer le mandant et en demander
-        l’autorisation ;</li>
-      <li>mettre à jour puis à supprimer l’ensemble des informations relatives au mandant lorsqu’elles ne sont plus
-        nécessaires à la réalisation des démarches lui incombant au titre du mandat ;</li>
-      <li>informer le mandant de toutes les actions qu’il/elle effectue à sa place ;</li>
-      <li>conserver les données à caractère personnel strictement nécessaires à ces démarches le seul temps strictement
-        nécessaire à leur réalisation.</li>
+      <li>ne collecter et conserver que les informations strictement nécessaires aux démarches listées,</li>
+      <li>n’utiliser les informations fournies que pour les démarches listées. Pour effectuer d’autres démarches, il ou elle doit au préalable en informer l’usager et demander l’autorisation,</li>
+      <li>ne collecter et conserver que les informations strictement nécessaires aux périmètres de démarches listés,</li>
+      <li>n’utiliser les informations fournies que pour la réalisation de ces démarches. Pour effectuer d’autres démarches, il ou elle doit au préalable en informer l’usager et demander l’autorisation,</li>
+      <li>cmettre à jour puis à supprimer l’ensemble des informations relatives à l’usager lorsqu’elles ne sont plus nécessaires à la réalisation des démarches,</li>
+      <li>respecter la confidentialité des informations fournies,</li>
+      <li>prendre toutes les précautions pour assurer la sécurité des données personnelles transmises.</li>
     </ol>
   </p>
 
   <p>
+    L’acceptation du mandat par le mandataire est réputée acquise par la réalisation effective des missions détaillées dans le présent mandat.
     Le mandataire est soumis à une obligation de confidentialité.
     Il ne doit en aucun cas divulguer les informations du mandant à des tiers lorsque cette divulgation n’est
-    pas nécessaire à l’accomplissement des démarches dont il est responsable (ex. : il ne doit pas
-    communiquer des informations concernant le mandant à son collègue de travail). Le mandataire enregistre les
+    pas nécessaire à l’accomplissement des démarches dont il est responsable. Le mandataire enregistre les
     informations du mandant de manière sécurisée et notamment prend toutes précautions pour assurer la sécurité physique
     et logique de ces données.
   </p>
 
   <p>
     Le présent mandat est accepté et consenti pour la durée nécessaire à l’accomplissement des missions de
-    <strong>{{ aidant }}</strong>.
+    l'aidant habilité de <strong>{{ organisation }}</strong>.
     Le mandat prend fin lorsque la réalisation des démarches susvisées a été accomplie, ou à tout moment si le mandant
     ou le mandataire décide de révoquer le mandat.
   </p>
 
   <p>
-    <strong>{{ aidant }}</strong> est tenu(e) d’accomplir le mandat tant qu’il en demeure chargé, et répond des dommages
+    L'aidant habilité de <strong>{{ organisation }}</strong> est tenu d’accomplir le mandat tant qu’il en demeure chargé, et répond des dommages
     et intérêts qui
     pourraient résulter de son inexécution conformément à l’article 1991 du Code civil.
   </p>
@@ -86,6 +83,5 @@
 
   <p class="margin-bottom-50">
     <span>Le mandant</span>
-    <span class="margin-left-100">Le mandataire</span>
   </p>
 </section>
